+++ nengo_dl/__init__.py
@@ -22,8 +22,8 @@ if not any(["gpu" in x.name for x in device_lib.list_local_devices()]):
 # check nengo version
 from nengo.version import version_info as nengo_version
 
-minimum_nengo_version = (2, 3, 1)
-latest_nengo_version = (2, 4, 1)
+minimum_nengo_version = (2, 5, 0)
+latest_nengo_version = (2, 5, 1)
 if nengo_version < minimum_nengo_version:  # pragma: no cover
     raise ValueError(
         "`nengo_dl` does not support `nengo` version %s. Upgrade "